fix(types): expand sprite_slice dimensions to uint16_t to support large atlases

+++ include/scyan/scyan_assets.install.h
@@ -13,10 +13,10 @@
 
 /* Sprite slice descriptor */
 struct sprite_slice {
-    uint8_t x;
-    uint8_t y;
-    uint8_t width;
-    uint8_t height;
+    uint16_t x;
+    uint16_t y;
+    uint16_t width;
+    uint16_t height;
 };
 
 /* Symbol identifiers */
